% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MsBackendTimsTof-functions.R
\name{setup_converter_library}
\alias{setup_converter_library}
\title{Function to setup the required built-in open-source converters or the Bruker
library to convert tof-to-mz and scan-to-inv_ion_mobility.
When the Burker option is selected the library is downloaded automatically
using \code{opentimsr::download_bruker_proprietary_code()}.}
\usage{
setup_converter_library(opensource = TRUE, path = NULL, force = FALSE)
}
\arguments{
\item{opensource}{\code{logical} determine if use the built-in open-source
converters of \code{opentimsr} (defaults) or the proprietary library from
Bruker.}

\item{path}{\code{character(1)} where save the Bruker library. If \code{NULL} the
library is cached using \code{BiocFileCache}. Used only with the Bruker
library.}

\item{force}{\code{logical} force to redownload the Bruker library (default:
\code{FALSE}).}
}
\description{
Function to setup the required built-in open-source converters or the Bruker
library to convert tof-to-mz and scan-to-inv_ion_mobility.
When the Burker option is selected the library is downloaded automatically
using \code{opentimsr::download_bruker_proprietary_code()}.
}
\examples{

## To setup the open-source built-in library
setup_converter_library()

## To use the Bruker library and cache it
setup_converter_library(opensource = FALSE)

## To use the Bruker library and save it in a personal folder
setup_converter_library(opensource = FALSE, path = tempdir())

}
\author{
Gabriele Tomè
}
